+++ lib/visibility_aware_state.dart
@@ -6,11 +6,10 @@ import 'package:visibility_aware_state/extensions.dart';
 
 abstract class VisibilityAwareState<T extends StatefulWidget> extends State<T>
 // ignore: prefer_mixin
-    with WidgetsBindingObserver, _StackChangedListener {
-
-  VisibilityAwareState({
-    this.debugPrintsEnabled = false
-  });
+    with
+        WidgetsBindingObserver,
+        _StackChangedListener {
+  VisibilityAwareState({this.debugPrintsEnabled = false});
 
   static final Set<String> _widgetStack = {};
   static final Map<String, int> _widgetStackTimestamps = {};
@@ -30,7 +29,8 @@ abstract class VisibilityAwareState<T extends StatefulWidget> extends State<T>
   static bool _addToStack(String widgetName) {
     final bool result = _widgetStack.add(widgetName);
     if (result) {
-      _widgetStackTimestamps[widgetName] = DateTime.now().millisecondsSinceEpoch;
+      _widgetStackTimestamps[widgetName] =
+          DateTime.now().millisecondsSinceEpoch;
       for (final listener in _listeners) {
         listener._onAddToStack(widgetName);
       }
@@ -57,14 +57,18 @@ abstract class VisibilityAwareState<T extends StatefulWidget> extends State<T>
 
   @override
   void _onAddToStack(String widgetName) {
-    if (_widgetVisibility != WidgetVisibility.INVISIBLE && runtimeType.toString() != widgetName && !_wasAddedTogetherWith(widgetName)) {
+    if (_widgetVisibility != WidgetVisibility.INVISIBLE &&
+        runtimeType.toString() != widgetName &&
+        !_wasAddedTogetherWith(widgetName)) {
       _onVisibilityChanged(WidgetVisibility.INVISIBLE);
     }
   }
 
   @override
   void _onRemoveFromStack() {
-    if (runtimeType.toString() == _widgetStack.last || _wasAddedTogetherWith(_widgetStack.last)) {
+    if (_widgetStack.isNotEmpty &&
+        (runtimeType.toString() == _widgetStack.last ||
+            _wasAddedTogetherWith(_widgetStack.last))) {
       _onVisibilityChanged(WidgetVisibility.VISIBLE);
     }
   }
@@ -117,7 +121,8 @@ abstract class VisibilityAwareState<T extends StatefulWidget> extends State<T>
       _onVisibilityChanged(WidgetVisibility.INVISIBLE);
     } else if (state == AppLifecycleState.resumed) {
       // user returned to our app
-      if (runtimeType.toString() == _widgetStack.last || _wasAddedTogetherWith(_widgetStack.last)) {
+      if (runtimeType.toString() == _widgetStack.last ||
+          _wasAddedTogetherWith(_widgetStack.last)) {
         _onVisibilityChanged(WidgetVisibility.VISIBLE);
       }
     } else if (state == AppLifecycleState.detached) {
@@ -138,7 +143,8 @@ abstract class VisibilityAwareState<T extends StatefulWidget> extends State<T>
 
     if (diff < 50) {
       if (debugPrintsEnabled) {
-        debugPrint('diff of $otherWidgetsName and ${runtimeType.toString()}: $diff');
+        debugPrint(
+            'diff of $otherWidgetsName and ${runtimeType.toString()}: $diff');
       }
       return true;
     }
@@ -151,8 +157,8 @@ abstract class VisibilityAwareState<T extends StatefulWidget> extends State<T>
       onVisibilityChanged(visibility);
 
       if (debugPrintsEnabled) {
-        debugPrint('$runtimeType.onVisibilityChanged(${enumToString(
-            visibility)}) - $_widgetStack');
+        debugPrint(
+            '$runtimeType.onVisibilityChanged(${enumToString(visibility)}) - $_widgetStack');
       }
     }
   }
@@ -173,16 +179,10 @@ abstract class VisibilityAwareState<T extends StatefulWidget> extends State<T>
   }
 }
 
-enum WidgetVisibility {
-  VISIBLE,
-  INVISIBLE,
-  GONE
-}
+enum WidgetVisibility { VISIBLE, INVISIBLE, GONE }
 
 mixin _StackChangedListener {
-
   void _onAddToStack(String widgetName);
 
   void _onRemoveFromStack();
-
 }